requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
flask-restplus==0.13.0
flask-restx==0.2.0
fuzzywuzzy==0.18.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.16.0
python-Levenshtein==0.12.0
pytz==2020.1
six==1.14.0
Werkzeug==1.0.1
zipp==3.1.0
boto3
datetime
shortuuid
azure-ai-textanalytics==5.0.0
azure-common==1.1.25
azure-core==1.8.2
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.5.0